调整插件模板位置
