+++ server/initialize/plugin.go
@@ -2,9 +2,7 @@ package initialize
 
 import (
 	"github.com/flipped-aurora/gin-vue-admin/server/global"
-	email "github.com/flipped-aurora/gva-plugins/email" // 在线仓库模式go
-	//"github.com/flipped-aurora/gin-vue-admin/server/plugin/email" // 本地插件仓库地址模式
-	"github.com/flipped-aurora/gin-vue-admin/server/plugin/example_plugin"
+	"github.com/flipped-aurora/gin-vue-admin/server/plugin/email"
 	"github.com/flipped-aurora/gin-vue-admin/server/utils/plugin"
 	"github.com/gin-gonic/gin"
 )
@@ -17,10 +15,6 @@ func PluginInit(group *gin.RouterGroup, Plugin ...plugin.Plugin) {
 }
 
 func InstallPlugin(PublicGroup *gin.RouterGroup, PrivateGroup *gin.RouterGroup) {
-	//  添加开放权限的插件 示例
-	PluginInit(PublicGroup,
-		example_plugin.ExamplePlugin)
-
 	//  添加跟角色挂钩权限的插件 示例 本地示例模式于在线仓库模式注意上方的import 可以自行切换 效果相同
 	PluginInit(PrivateGroup, email.CreateEmailPlug(
 		global.GVA_CONFIG.Email.To,
